Converted AppHeaderLoginMain and AppHeaderMain components to .tsx file
