+++ components/custom_typings/image-loader.d.ts
@@ -1,0 +1,9 @@
+declare module "*.png" {
+  const value: any;
+  export default value;
+}
+
+declare module "*.svg" {
+  const value: any;
+  export default value;
+}
