+++ storybook/tsconfig.json
@@ -7,5 +7,8 @@
     "jsx": "react",
     "strict": false,
     "esModuleInterop": true
-  }
+  },
+  "files": [
+    "custom_typings/image-loader.d.ts"
+  ]
 }
